Update the correct files.
